[supercell] Port Kammerl Beat-Repeat Effect v0.31

+++ clouds/cv_scaler.h
@@ -101,7 +101,7 @@ class CvScaler {
   }
 
  private:
-  static const int8_t kAdcLatency = 5;
+  static const int kAdcLatency = 5;
   
   Adc adc_;
   PotsAdc pots_adc_;
